Applied all suggestions
- Removed unnecessary useStyles.js
- Refactored BookingCard and BookingForm
- Added helping comments
- Grouped MUI components
