+++ client/src/pages/Listings/Listings.tsx
@@ -1,14 +1,8 @@
-import Box from '@material-ui/core/Box';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
+import { Grid, Box, Typography } from '@mui/material';
 import { Link } from 'react-router-dom';
 import Layout from '../Layout/Layout';
-import useStyles from './useStyles';
 
 const Listings = (): JSX.Element => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const classes = useStyles();
-
   return (
     <Layout>
       <Grid>
