--- client/src/pages/Listings/useStyles.ts
@@ -1,5 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(() => ({}));
-
-export default useStyles;
