+++ client/src/pages/MySitters/MySitters.tsx
@@ -1,13 +1,7 @@
-import Box from '@material-ui/core/Box';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
+import { Grid, Box, Typography } from '@mui/material';
 import Layout from '../Layout/Layout';
-import useStyles from './useStyles';
 
 const MySitters = (): JSX.Element => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const classes = useStyles();
-
   return (
     <Layout>
       <Grid>
