+++ client/src/pages/TempDashboardMain/TempDashboardMain.tsx
@@ -2,12 +2,8 @@ import Box from '@material-ui/core/Box';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import Layout from '../Layout/Layout';
-import useStyles from './useStyles';
 
 const TempDashboardMain = (): JSX.Element => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const classes = useStyles();
-
   return (
     <Layout>
       <Grid>
